staging: sync: Use an on-stack allocation for fence info ioctl
Since the fence info ioctl limits output data length to 4096 bytes, we
can just use a 4096-byte on-stack buffer for it (which is confirmed to
be safe via runtime stack usage measurements). This ioctl is used for
every frame rendered to the display, so eliminating dynamic memory
allocation overhead here improves frame rendering performance.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Ratoriku <a1063021545@gmail.com>
Signed-off-by: Moeeze Hassan <fammas.maz@gmail.com>