sched: bias tasks to gold without ta boost
Gold Cluster is much more efficient and faster compared to the Silver, since we are using higher freqs by default, the tasks can benefit from being completed faster and putting cpu back to idle. To bias, usually top-app stune boost is set to a value higher than 1; creating an artificial load of 1% percent, and biasing to big cluster. It was observed that a stune boost of more than 1 was infact more beneficial for the battery. Hence, try to bias tasks to big cluster by defaul even when there is no TA boost.

Signed-off-by: Moeeze Hassan <fammas.maz@gmail.com>