sched/tune: only force UX tasks to big cores
If we force top-app's boost value to be >0 we get excellent
jitter values due to top-app being rendered on big cores.

However along with the actual top app a load of crap ends up on
big cluster too. We can avoid this by checking if adj_score is 0,
which means that this app is actually the one that's currently on
the screen (or it's one of its child processes).

Signed-off-by: Yaroslav Furman <yaro330@gmail.com>
(cherry picked from commit 1614cdb4814c58a0c42ac9dca1ce887436f494ab)

Change-Id: Iad7abce83d6946e53b5da8c9cdd44f9a39bda832
Signed-off-by: Moeeze Hassan <fammas.maz@gmail.com>